+++ Assets/AppCenter/Editor/AndroidGradleHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5498966cb5c28e94ebab8714275d6ccb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
